+++ r/pilotr/examples/run_demo.R
@@ -5,6 +5,7 @@ args <- commandArgs(trailingOnly = FALSE)
 here <- dirname(normalizePath(sub("^--file=", "", args[grep("^--file=", args)])))
 source(file.path(here, "..", "R", "core.R"))
 source(file.path(here, "..", "R", "simulate.R"))
+source(file.path(here, "..", "R", "parallel.R"))
 source(file.path(here, "..", "R", "power.R"))
 
 SPEC  <- file.path(here, "..", "..", "..", "spec", "examples")